Update `Updater` related methods
1. Added `sectionReaderWriter` to make `NewUpdater` create new
`Updater` from `io.ReadWriteSeeker` instead of `os.File`.
2. Updated to permit `Updater` overwrite existing file data.
3. Keep `zip.Reader` and `zip.Writer` source code same with go 1.21.3 .

+++ writer.go
@@ -1,3 +1,7 @@
+// Copyright 2011 The Go Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style
+// license that can be found in the LICENSE file.
+
 package zip
 
 import (
@@ -32,7 +36,7 @@ type Writer struct {
 
 type header struct {
 	*FileHeader
-	offset int64
+	offset uint64
 	raw    bool
 }
 
@@ -109,7 +113,7 @@ func (w *Writer) Close() error {
 			eb.uint16(24) // size = 3x uint64
 			eb.uint64(h.UncompressedSize64)
 			eb.uint64(h.CompressedSize64)
-			eb.uint64(uint64(h.offset))
+			eb.uint64(h.offset)
 			h.Extra = append(h.Extra, buf[:]...)
 		} else {
 			b.uint32(h.CompressedSize)
@@ -331,7 +335,7 @@ func (w *Writer) CreateHeader(fh *FileHeader) (io.Writer, error) {
 	)
 	h := &header{
 		FileHeader: fh,
-		offset:     w.cw.count,
+		offset:     uint64(w.cw.count),
 	}
 
 	if strings.HasSuffix(fh.Name, "/") {
@@ -445,7 +449,7 @@ func (w *Writer) CreateRaw(fh *FileHeader) (io.Writer, error) {
 
 	h := &header{
 		FileHeader: fh,
-		offset:     w.cw.count,
+		offset:     uint64(w.cw.count),
 		raw:        true,
 	}
 	w.dir = append(w.dir, h)
